casting issue + corrected comment

+++ protocol/server/audio.c
@@ -134,7 +134,7 @@ int32_t multithreaded_send_audio(void* opaque) {
                         audio_frame_size = 0;
                     }
 
-                    if (audio_encoder->encoded_frame_size > MAX_AUDIOFRAME_DATA_SIZE) {
+                    if (audio_encoder->encoded_frame_size > (int)MAX_AUDIOFRAME_DATA_SIZE) {
                         LOG_ERROR("Audio data too large: %d", audio_encoder->encoded_frame_size);
                     } else {
                         static char buf[LARGEST_AUDIOFRAME_SIZE];
